# be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/event-groups": {
      "post": {
        "operationId": "EventGroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event group created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create event group",
        "tags": [
          "Event Groups"
        ]
      },
      "get": {
        "operationId": "EventGroupsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of event groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventGroupResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List event groups",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}": {
      "get": {
        "operationId": "EventGroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event group found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get event group by ID",
        "tags": [
          "Event Groups"
        ]
      },
      "patch": {
        "operationId": "EventGroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update event group",
        "tags": [
          "Event Groups"
        ]
      },
      "delete": {
        "operationId": "EventGroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event group deleted"
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete event group",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}/periods": {
      "post": {
        "operationId": "EventGroupsController_createPeriod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Period created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create period for event group",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}/details": {
      "get": {
        "operationId": "EventGroupsController_getDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupDetailsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get event group details with analytics",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/periods/{id}": {
      "patch": {
        "operationId": "EventPeriodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event period UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventPeriodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update event period",
        "tags": [
          "Event Periods"
        ]
      },
      "delete": {
        "operationId": "EventPeriodsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event period UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Period deleted"
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete event period",
        "tags": [
          "Event Periods"
        ]
      }
    },
    "/api/v1/periods/{id}/close": {
      "post": {
        "operationId": "EventPeriodsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event period UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseEventPeriodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Period closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close an active event period",
        "tags": [
          "Event Periods"
        ]
      }
    },
...
    "/api/v1/timeline": {
      "get": {
        "operationId": "TimelineController_getTimeline",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Start of timeline range (defaults to 1 year ago)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "End of timeline range (defaults to today)",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vertical timeline (events + days for a date range)",
        "tags": [
          "Timeline"
        ]
      }
    },
    "/api/v1/timeline/week": {
      "get": {
        "operationId": "TimelineController_getWeekTimeline",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Any date within the target week (YYYY-MM-DD)",
            "schema": {
              "example": "2024-06-15",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekTimelineResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get week view (7 days with events for a specific week)",
        "tags": [
          "Timeline"
        ]
      }
    },
...
  },
  "info": {
    "title": "LifeSpan API",
    "description": "Backend API for Life Timeline — build a visual timeline of your life",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEventGroupDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "categoryId",
          "title"
        ]
      },
      "EventGroupCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "EventPeriodResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "example": "2024-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2024-01-31",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "startDate",
          "createdAt"
        ]
      },
      "EventGroupResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventGroupCategoryDto"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPeriodResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "periods",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateEventGroupDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateEventPeriodDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "startDate"
        ]
      },
      "MoodStatDto": {
        "type": "object",
        "properties": {
          "dayStateName": {
            "type": "string"
          },
          "dayStateColor": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          }
        },
        "required": [
          "dayStateName",
          "dayStateColor",
          "count",
          "percentage"
        ]
      },
      "EventGroupDetailsMediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "s3Key",
          "url",
          "fileName",
          "contentType",
          "size",
          "createdAt"
        ]
      },
      "MoodDistributionItemDto": {
        "type": "object",
        "properties": {
          "moodId": {
            "type": "string"
          },
          "moodName": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          }
        },
        "required": [
          "moodId",
          "moodName",
          "color",
          "count",
          "percentage"
        ]
      },
      "PeriodDensityDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "example": "2024-01-01"
          },
          "end": {
            "type": "string",
            "example": "2024-01-31"
          },
          "activeDays": {
            "type": "number"
          }
        },
        "required": [
          "start",
          "end",
          "activeDays"
        ]
      },
      "ChapterAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalPeriods": {
            "type": "number"
          },
          "totalDays": {
            "type": "number"
          },
          "totalMedia": {
            "type": "number"
          },
          "averageMoodScore": {
            "type": "number",
            "nullable": true
          },
          "moodDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoodDistributionItemDto"
            }
          },
          "density": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodDensityDto"
            }
          }
        },
        "required": [
          "totalPeriods",
          "totalDays",
          "totalMedia",
          "moodDistribution",
          "density"
        ]
      },
      "EventGroupDetailsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventGroupCategoryDto"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPeriodResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "moodStats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoodStatDto"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventGroupDetailsMediaDto"
            }
          },
          "totalDays": {
            "type": "number"
          },
          "analytics": {
            "$ref": "#/components/schemas/ChapterAnalyticsDto"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "periods",
          "createdAt",
          "updatedAt",
          "moodStats",
          "media",
          "totalDays",
          "analytics"
        ]
      },
      "UpdateEventPeriodDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "CloseEventPeriodDto": {
        "type": "object",
        "properties": {
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "endDate"
        ]
      },
...
      "TimelineEventGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "TimelineCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "TimelinePeriodDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "eventGroup": {
            "$ref": "#/components/schemas/TimelineEventGroupDto"
          },
          "category": {
            "$ref": "#/components/schemas/TimelineCategoryDto"
          }
        },
        "required": [
          "id",
          "startDate",
          "endDate",
          "comment",
          "eventGroup",
          "category"
        ]
      },
      "TimelineDayStateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "TimelineMediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "s3Key",
          "url",
          "fileName",
          "contentType",
          "size",
          "createdAt"
        ]
      },
      "TimelineDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2024-01-15"
          },
          "dayState": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TimelineDayStateDto"
              }
            ]
          },
          "mainMediaId": {
            "type": "string",
            "nullable": true,
            "example": "media-uuid",
            "description": "Cover media ID"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineMediaDto"
            }
          }
        },
        "required": [
          "date",
          "dayState",
          "mainMediaId",
          "media"
        ]
      },
      "TimelineResponseDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "example": "2024-01-01"
          },
          "to": {
            "type": "string",
            "example": "2024-12-31"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelinePeriodDto"
            }
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineDayDto"
            }
          }
        },
        "required": [
          "from",
          "to",
          "periods",
          "days"
        ]
      },
      "WeekTimelineResponseDto": {
        "type": "object",
        "properties": {
          "weekStart": {
            "type": "string",
            "example": "2024-06-10",
            "description": "Monday of the week"
          },
          "weekEnd": {
            "type": "string",
            "example": "2024-06-16",
            "description": "Sunday of the week"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelinePeriodDto"
            }
          },
          "days": {
            "description": "Exactly 7 days (Mon-Sun), with or without state",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineDayDto"
            }
          }
        },
        "required": [
          "weekStart",
          "weekEnd",
          "periods",
          "days"
        ]
      },
...
    }
  }
}